+++ src/index.js
@@ -1,15 +1,22 @@
 import path from 'path';
+import os from 'os';
+import crypto from 'crypto';
 
 import webpack from 'webpack';
 import validateOptions from 'schema-utils';
 import pLimit from 'p-limit';
-
 import globby from 'globby';
+import findCacheDir from 'find-cache-dir';
+import serialize from 'serialize-javascript';
+import cacache from 'cacache';
+import loaderUtils from 'loader-utils';
+import normalizePath from 'normalize-path';
+
+import { version } from '../package.json';
 
 import schema from './options.json';
-import postProcessPattern from './postProcessPattern';
 import isTemplateLike from './utils/isTemplateLike';
-import { stat } from './utils/promisify';
+import { readFile, stat } from './utils/promisify';
 import createPatternGlob from './utils/createPatternGlob';
 
 // webpack 5 exposes the sources property to ensure the right version of webpack-sources is used
@@ -167,7 +174,131 @@ class CopyPlugin {
     });
 
     return Promise.all(
-      files.map((file) => postProcessPattern(globalRef, pattern, file))
+      files.map(async (file) => {
+        // If this came from a glob, add it to the file watchlist
+        if (pattern.fromType === 'glob') {
+          logger.debug(`add ${file.absoluteFrom} as fileDependencies`);
+
+          globalRef.compilation.fileDependencies.add(file.absoluteFrom);
+        }
+
+        logger.debug(`reading '${file.absoluteFrom}' to write to assets`);
+
+        let data;
+
+        try {
+          data = await readFile(globalRef.inputFileSystem, file.absoluteFrom);
+        } catch (error) {
+          globalRef.compilation.errors.push(error);
+
+          return;
+        }
+
+        if (pattern.transform) {
+          logger.log(`transforming content for '${file.absoluteFrom}'`);
+
+          if (pattern.cacheTransform) {
+            const cacheDirectory = pattern.cacheTransform.directory
+              ? pattern.cacheTransform.directory
+              : typeof pattern.cacheTransform === 'string'
+              ? pattern.cacheTransform
+              : findCacheDir({ name: 'copy-webpack-plugin' }) || os.tmpdir();
+            let defaultCacheKeys = {
+              version,
+              transform: pattern.transform,
+              contentHash: crypto.createHash('md4').update(data).digest('hex'),
+            };
+
+            if (typeof pattern.cacheTransform.keys === 'function') {
+              defaultCacheKeys = await pattern.cacheTransform.keys(
+                defaultCacheKeys,
+                file.absoluteFrom
+              );
+            } else {
+              defaultCacheKeys = {
+                ...defaultCacheKeys,
+                ...pattern.cacheTransform.keys,
+              };
+            }
+
+            const cacheKeys = serialize(defaultCacheKeys);
+
+            try {
+              const result = await cacache.get(cacheDirectory, cacheKeys);
+
+              logger.debug(
+                `getting cached transformation for '${file.absoluteFrom}'`
+              );
+
+              ({ data } = result);
+            } catch (_ignoreError) {
+              data = await pattern.transform(data, file.absoluteFrom);
+
+              logger.debug(`caching transformation for '${file.absoluteFrom}'`);
+
+              await cacache.put(cacheDirectory, cacheKeys, data);
+            }
+          } else {
+            data = await pattern.transform(data, file.absoluteFrom);
+          }
+        }
+
+        if (pattern.toType === 'template') {
+          logger.log(
+            `interpolating template '${file.webpackTo}' for '${file.relativeFrom}'`
+          );
+
+          // If it doesn't have an extension, remove it from the pattern
+          // ie. [name].[ext] or [name][ext] both become [name]
+          if (!path.extname(file.relativeFrom)) {
+            // eslint-disable-next-line no-param-reassign
+            file.webpackTo = file.webpackTo.replace(/\.?\[ext]/g, '');
+          }
+
+          // eslint-disable-next-line no-param-reassign
+          file.immutable = /\[(?:([^:\]]+):)?(?:hash|contenthash)(?::([a-z]+\d*))?(?::(\d+))?\]/gi.test(
+            file.webpackTo
+          );
+
+          // eslint-disable-next-line no-param-reassign
+          file.webpackTo = loaderUtils.interpolateName(
+            { resourcePath: file.absoluteFrom },
+            file.webpackTo,
+            {
+              content: data,
+              context: pattern.context,
+            }
+          );
+
+          // Bug in `loader-utils`, package convert `\\` to `/`, need fix in loader-utils
+          // eslint-disable-next-line no-param-reassign
+          file.webpackTo = path.normalize(file.webpackTo);
+        }
+
+        if (pattern.transformPath) {
+          logger.log(
+            `transforming path '${file.webpackTo}' for '${file.absoluteFrom}'`
+          );
+
+          // eslint-disable-next-line no-param-reassign
+          file.immutable = false;
+          // eslint-disable-next-line no-param-reassign
+          file.webpackTo = await pattern.transformPath(
+            file.webpackTo,
+            file.absoluteFrom
+          );
+        }
+
+        // eslint-disable-next-line no-param-reassign
+        file.data = data;
+        // eslint-disable-next-line no-param-reassign
+        file.targetPath = normalizePath(file.webpackTo);
+        // eslint-disable-next-line no-param-reassign
+        file.force = pattern.force;
+
+        // eslint-disable-next-line consistent-return
+        return file;
+      })
     );
   }
